Implement hook to fetch dataset definition from remote source
- Move existing/internal/local dataset definition to new file:
  `datasets.js`.
- Implement new `useDatasets()` hook that fetches dataset definition
  from JSON file at remote location.
- Remote location containing dataset definition is set through
  `REACT_APP_URL_DATASETS` environment variable. If that variable is not
  set or loading failed, e.g. because the target file does not exist,
  vitessce will fall back to internal datasets (`datasets.js`).
- A "Loading..." panel is shown while data is being fetched from remote
  source.
- A new `<App>` component is defined to use the `useDatasets()` hook
  within and then decide whether to show a "Loading..." panel, the
  `<Welcome>` component or the <Vitessce>` component (if a dataset ID is
  provided in the URL). The `createApp()` function now returns the
  `<App>` component.
- The `getConfig()` and `listConfigs()` now take the available datasets
  as an input parameter.
